+++ admin/includes/view_all_users.php
@@ -78,33 +78,33 @@
 </table>
 
 <?php
-    if(isset($_GET['change_to_admin']))
+    if(isset($_SESSION['user_role']))
     {
-        $get_user_id = $_GET['change_to_admin'];
-        $query = "UPDATE users SET user_role = 'admin' WHERE user_id = $get_user_id";
+        if($_SESSION['user_role'] == 'admin')
+        {
+            if(isset($_GET['change_to_admin']))
+            {           
+                $get_user_id = mysqli_real_escape_string($connection,$_GET['change_to_admin']);
+                $query = "UPDATE users SET user_role = 'admin' WHERE user_id = $get_user_id";
+                $change_to_admin_query = mysqli_query($connection, $query);
+                header("Location: users.php");
+            }    
 
-        $change_to_admin_query = mysqli_query($connection, $query);
+            if(isset($_GET['change_to_sub']))
+            {
+                $get_user_id = mysqli_real_escape_string($connection,$_GET['change_to_sub']);
+                $query = "UPDATE users SET user_role = 'subscriber' WHERE user_id = $get_user_id";
+                $change_to_sub_query = mysqli_query($connection, $query);
+                header("Location: users.php");                    
+            }
 
-        header("Location: users.php");
-    }
-
-    if(isset($_GET['change_to_sub']))
-    {
-        $get_user_id = $_GET['change_to_sub'];
-        $query = "UPDATE users SET user_role = 'subscriber' WHERE user_id = $get_user_id";
-
-        $change_to_sub_query = mysqli_query($connection, $query);
-
-        header("Location: users.php");
-    }
-
-    if(isset($_GET['delete']))
-    {
-        $get_user_id = $_GET['delete'];
-        $query = "DELETE FROM users WHERE user_id = {$get_user_id}";
-        
-        $delete_query = mysqli_query($connection, $query);
-        
-        header("Location: users.php");
+            if(isset($_GET['delete']))
+            {
+                $get_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
+                $query = "DELETE FROM users WHERE user_id = {$get_user_id}";      
+                $delete_query = mysqli_query($connection, $query);    
+                header("Location: users.php");
+            }
+        }
     }
 ?>
